fix: duplicate login+consent in didcomm flows
This is a PoC patch to enable the following:

* first login at Issuer creates a session stored in your browser.
If this session is detected then the login form will be skipped
entirely.
* the concept of "trusted domains" is introduced in the login-consent
app. This allows the app to automatically accept consent requests for
certain domains. This can be used to skip the consent request for the
issuer domain within which the login-consent app is hosted

Signed-off-by: George Aristy <george.aristy@securekey.com>